Refactoring of conformance tests  (#2771)
