+++ activiti-spring-conformance-tests/activiti-spring-conformance-set0/src/test/java/org/activiti/spring/conformance/set0/ConformanceBasicProcessInformationTest.java
@@ -1,6 +1,5 @@
 package org.activiti.spring.conformance.set0;
 
-import static org.activiti.spring.conformance.set0.Set0RuntimeTestConfiguration.collectedEvents;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.catchThrowable;
 
@@ -13,6 +12,7 @@ import org.activiti.api.process.model.events.BPMNSequenceFlowTakenEvent;
 import org.activiti.api.process.model.events.ProcessRuntimeEvent;
 import org.activiti.api.process.runtime.ProcessRuntime;
 import org.activiti.api.runtime.shared.NotFoundException;
+import org.activiti.spring.conformance.util.RuntimeTestConfiguration;
 import org.activiti.spring.conformance.util.security.SecurityUtil;
 import org.junit.After;
 import org.junit.Test;
@@ -34,7 +34,7 @@ public class ConformanceBasicProcessInformationTest {
 
     @After
     public void cleanUp() {
-        collectedEvents.clear();
+        RuntimeTestConfiguration.collectedEvents.clear();
     }
 
     /*
@@ -74,7 +74,7 @@ public class ConformanceBasicProcessInformationTest {
         assertThat(throwable)
                 .isInstanceOf(NotFoundException.class);
 
-        assertThat(collectedEvents).extracting(RuntimeEvent::getEventType).containsExactly(
+        assertThat(RuntimeTestConfiguration.collectedEvents).extracting(RuntimeEvent::getEventType).containsExactly(
                 ProcessRuntimeEvent.ProcessEvents.PROCESS_CREATED,
                 ProcessRuntimeEvent.ProcessEvents.PROCESS_STARTED,
                 BPMNActivityEvent.ActivityEvents.ACTIVITY_STARTED,
@@ -136,7 +136,7 @@ public class ConformanceBasicProcessInformationTest {
         assertThat(throwable)
                 .isInstanceOf(NotFoundException.class);
 
-        assertThat(collectedEvents).extracting(RuntimeEvent::getEventType).containsExactly(
+        assertThat(RuntimeTestConfiguration.collectedEvents).extracting(RuntimeEvent::getEventType).containsExactly(
                 ProcessRuntimeEvent.ProcessEvents.PROCESS_CREATED,
                 VariableEvent.VariableEvents.VARIABLE_CREATED,
                 ProcessRuntimeEvent.ProcessEvents.PROCESS_STARTED,
