+++ activiti-spring-conformance-tests/activiti-spring-conformance-set2/src/test/java/org/activiti/spring/conformance/set2/UserTaskAssigneeDeleteRuntimeTest.java
@@ -1,6 +1,5 @@
 package org.activiti.spring.conformance.set2;
 
-import static org.activiti.spring.conformance.set2.Set2RuntimeTestConfiguration.collectedEvents;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.catchThrowable;
 
@@ -19,6 +18,7 @@ import org.activiti.api.task.model.builders.TaskPayloadBuilder;
 import org.activiti.api.task.model.events.TaskRuntimeEvent;
 import org.activiti.api.task.runtime.TaskRuntime;
 import org.activiti.engine.ActivitiException;
+import org.activiti.spring.conformance.util.RuntimeTestConfiguration;
 import org.activiti.spring.conformance.util.security.SecurityUtil;
 import org.junit.After;
 import org.junit.Before;
@@ -48,7 +48,7 @@ public class UserTaskAssigneeDeleteRuntimeTest {
 
     @Before
     public void cleanUp() {
-        collectedEvents.clear();
+        clearEvents();
     }
 
 
@@ -90,7 +90,7 @@ public class UserTaskAssigneeDeleteRuntimeTest {
 
         assertThat(task.getAssignee()).isEqualTo("user1");
 
-        assertThat(collectedEvents)
+        assertThat(RuntimeTestConfiguration.collectedEvents)
                 .extracting(RuntimeEvent::getEventType)
                 .containsExactly(
                         ProcessRuntimeEvent.ProcessEvents.PROCESS_CREATED,
@@ -102,7 +102,7 @@ public class UserTaskAssigneeDeleteRuntimeTest {
                         TaskRuntimeEvent.TaskEvents.TASK_CREATED,
                         TaskRuntimeEvent.TaskEvents.TASK_ASSIGNED);
 
-        collectedEvents.clear();
+        clearEvents();
 
         Throwable throwable = catchThrowable(() ->  taskRuntime.delete(TaskPayloadBuilder.delete()
                 .withTaskId(task.getId())
@@ -122,6 +122,11 @@ public class UserTaskAssigneeDeleteRuntimeTest {
         for(ProcessInstance pi : processInstancePage.getContent()){
             processAdminRuntime.delete(ProcessPayloadBuilder.delete(pi.getId()));
         }
+        clearEvents();
+    }
+    
+    public void clearEvents() {
+        RuntimeTestConfiguration.collectedEvents.clear();
     }
 
 }
