+++ activiti-spring-conformance-tests/activiti-spring-conformance-set5/src/test/java/org/activiti/spring/conformance/set5/Set5RuntimeTestConfiguration.java
@@ -1,125 +1,11 @@
 package org.activiti.spring.conformance.set5;
 
-import org.activiti.api.model.shared.event.RuntimeEvent;
-import org.activiti.api.model.shared.event.VariableCreatedEvent;
-import org.activiti.api.model.shared.event.VariableDeletedEvent;
-import org.activiti.api.model.shared.event.VariableUpdatedEvent;
-import org.activiti.api.process.model.events.BPMNActivityCancelledEvent;
-import org.activiti.api.process.model.events.BPMNActivityCompletedEvent;
-import org.activiti.api.process.model.events.BPMNActivityStartedEvent;
-import org.activiti.api.process.model.events.BPMNSequenceFlowTakenEvent;
 import org.activiti.api.process.runtime.connector.Connector;
-import org.activiti.api.process.runtime.events.*;
-import org.activiti.api.process.runtime.events.listener.BPMNElementEventListener;
-import org.activiti.api.process.runtime.events.listener.ProcessRuntimeEventListener;
-import org.activiti.api.runtime.shared.events.VariableEventListener;
-import org.activiti.api.task.runtime.events.*;
-import org.activiti.api.task.runtime.events.listener.TaskEventListener;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import java.util.ArrayList;
-import java.util.List;
-
 @Configuration
 public class Set5RuntimeTestConfiguration {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(Set5RuntimeTestConfiguration.class);
-
-    public static List<RuntimeEvent> collectedEvents = new ArrayList<>();
-
-
-    @Bean
-    public BPMNElementEventListener<BPMNActivityStartedEvent> bpmnActivityStartedListener() {
-        return bpmnActivityStartedEvent -> collectedEvents.add(bpmnActivityStartedEvent);
-    }
-
-    @Bean
-    public BPMNElementEventListener<BPMNActivityCompletedEvent> bpmnActivityCompletedListener() {
-        return bpmnActivityCompletedEvent -> collectedEvents.add(bpmnActivityCompletedEvent);
-    }
-
-    @Bean
-    public BPMNElementEventListener<BPMNActivityCancelledEvent> bpmnActivityCancelledListener() {
-        return bpmnActivityCancelledEvent -> collectedEvents.add(bpmnActivityCancelledEvent);
-    }
-
-    @Bean
-    public BPMNElementEventListener<BPMNSequenceFlowTakenEvent> bpmnSequenceFlowTakenListener() {
-        return bpmnSequenceFlowTakenEvent -> collectedEvents.add(bpmnSequenceFlowTakenEvent);
-    }
-
-    @Bean
-    public ProcessRuntimeEventListener<ProcessCreatedEvent> processCreatedListener() {
-        return processCreatedEvent -> collectedEvents.add(processCreatedEvent);
-    }
-
-    @Bean
-    public ProcessRuntimeEventListener<ProcessStartedEvent> processStartedListener() {
-        return processStartedEvent -> collectedEvents.add(processStartedEvent);
-    }
-
-    @Bean
-    public ProcessRuntimeEventListener<ProcessCompletedEvent> processCompletedListener() {
-        return processCompletedEvent -> collectedEvents.add(processCompletedEvent);
-    }
-
-    @Bean
-    public ProcessRuntimeEventListener<ProcessResumedEvent> processResumedListener() {
-        return processResumedEvent -> collectedEvents.add(processResumedEvent);
-    }
-
-    @Bean
-    public ProcessRuntimeEventListener<ProcessSuspendedEvent> processSuspendedListener() {
-        return processSuspendedEvent -> collectedEvents.add(processSuspendedEvent);
-    }
-
-    @Bean
-    public ProcessRuntimeEventListener<ProcessCancelledEvent> processCancelledListener() {
-        return processCancelledEvent -> collectedEvents.add(processCancelledEvent);
-    }
-
-    @Bean
-    public VariableEventListener<VariableCreatedEvent> variableCreatedEventListener() {
-        return variableCreatedEvent -> collectedEvents.add(variableCreatedEvent);
-    }
-
-    @Bean
-    public VariableEventListener<VariableDeletedEvent> variableDeletedEventListener() {
-        return variableDeletedEvent -> collectedEvents.add(variableDeletedEvent);
-    }
-
-    @Bean
-    public VariableEventListener<VariableUpdatedEvent> variableUpdatedEventListener() {
-        return variableUpdatedEvent -> collectedEvents.add(variableUpdatedEvent);
-    }
-
-    @Bean
-    public TaskEventListener<TaskCreatedEvent> taskCreatedEventListener() {
-        return taskCreatedEvent -> collectedEvents.add(taskCreatedEvent);
-    }
-
-    @Bean
-    public TaskEventListener<TaskUpdatedEvent> taskUpdatedEventListener() {
-        return taskUpdatedEvent -> collectedEvents.add(taskUpdatedEvent);
-    }
-
-    @Bean
-    public TaskEventListener<TaskCompletedEvent> taskCompletedEventListener() {
-        return taskCompletedEvent -> collectedEvents.add(taskCompletedEvent);
-    }
-
-    @Bean
-    public TaskEventListener<TaskSuspendedEvent> taskSuspendedEventListener() {
-        return taskSuspendedEvent -> collectedEvents.add(taskSuspendedEvent);
-    }
-
-    @Bean
-    public TaskEventListener<TaskAssignedEvent> taskAssignedEventListener() {
-        return taskAssignedEvent -> collectedEvents.add(taskAssignedEvent);
-    }
 
     @Bean(name = "connector-a")
     public Connector connectorA() {
