+++ activiti-spring-conformance-tests/activiti-spring-conformance-util/src/main/java/org/activiti/spring/conformance/util/RuntimeTestConfiguration.java
@@ -1,11 +1,36 @@
 package org.activiti.spring.conformance.util;
 
 import org.activiti.api.model.shared.event.RuntimeEvent;
+import org.activiti.api.model.shared.event.VariableCreatedEvent;
+import org.activiti.api.model.shared.event.VariableDeletedEvent;
+import org.activiti.api.model.shared.event.VariableUpdatedEvent;
+import org.activiti.api.process.model.events.BPMNActivityCancelledEvent;
+import org.activiti.api.process.model.events.BPMNActivityCompletedEvent;
+import org.activiti.api.process.model.events.BPMNActivityStartedEvent;
+import org.activiti.api.process.model.events.BPMNSequenceFlowTakenEvent;
+import org.activiti.api.process.model.events.BPMNSignalReceivedEvent;
+import org.activiti.api.process.runtime.events.ProcessCancelledEvent;
+import org.activiti.api.process.runtime.events.ProcessCompletedEvent;
+import org.activiti.api.process.runtime.events.ProcessCreatedEvent;
+import org.activiti.api.process.runtime.events.ProcessResumedEvent;
+import org.activiti.api.process.runtime.events.ProcessStartedEvent;
+import org.activiti.api.process.runtime.events.ProcessSuspendedEvent;
+import org.activiti.api.process.runtime.events.listener.BPMNElementEventListener;
+import org.activiti.api.process.runtime.events.listener.ProcessRuntimeEventListener;
+import org.activiti.api.runtime.shared.events.VariableEventListener;
+import org.activiti.api.task.runtime.events.TaskAssignedEvent;
+import org.activiti.api.task.runtime.events.TaskCancelledEvent;
+import org.activiti.api.task.runtime.events.TaskCompletedEvent;
+import org.activiti.api.task.runtime.events.TaskCreatedEvent;
+import org.activiti.api.task.runtime.events.TaskSuspendedEvent;
+import org.activiti.api.task.runtime.events.TaskUpdatedEvent;
+import org.activiti.api.task.runtime.events.listener.TaskEventListener;
 import org.activiti.core.common.spring.identity.ExtendedInMemoryUserDetailsManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
@@ -18,6 +43,8 @@ import java.util.List;
 public class RuntimeTestConfiguration {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(RuntimeTestConfiguration.class);
+    
+    public static List<RuntimeEvent> collectedEvents = new ArrayList<>();
 
     @Bean
     public UserDetailsService myUserDetailsService() {
@@ -65,6 +92,109 @@ public class RuntimeTestConfiguration {
 
 
         return extendedInMemoryUserDetailsManager;
+    }
+    
+    @Bean
+    public BPMNElementEventListener<BPMNActivityStartedEvent> bpmnActivityStartedListener() {
+        return bpmnActivityStartedEvent -> collectedEvents.add(bpmnActivityStartedEvent);
+    }
+
+    @Bean
+    public BPMNElementEventListener<BPMNActivityCompletedEvent> bpmnActivityCompletedListener() {
+        return bpmnActivityCompletedEvent -> collectedEvents.add(bpmnActivityCompletedEvent);
+    }
+
+    @Bean
+    public BPMNElementEventListener<BPMNActivityCancelledEvent> bpmnActivityCancelledListener() {
+        return bpmnActivityCancelledEvent -> collectedEvents.add(bpmnActivityCancelledEvent);
+    }
+
+
+    @Bean
+    public BPMNElementEventListener<BPMNSequenceFlowTakenEvent> bpmnSequenceFlowTakenListener() {
+        return bpmnSequenceFlowTakenEvent -> collectedEvents.add(bpmnSequenceFlowTakenEvent);
+    }
+
+    @Bean
+    public ProcessRuntimeEventListener<ProcessCreatedEvent> processCreatedListener() {
+        return processCreatedEvent -> collectedEvents.add(processCreatedEvent);
+    }
+
+    @Bean
+    public ProcessRuntimeEventListener<ProcessStartedEvent> processStartedListener() {
+        return processStartedEvent -> collectedEvents.add(processStartedEvent);
+    }
+
+    @Bean
+    public ProcessRuntimeEventListener<ProcessCompletedEvent> processCompletedListener() {
+        return processCompletedEvent -> collectedEvents.add(processCompletedEvent);
+    }
+
+    @Bean
+    public ProcessRuntimeEventListener<ProcessResumedEvent> processResumedListener() {
+        return processResumedEvent -> collectedEvents.add(processResumedEvent);
+    }
+
+    @Bean
+    public ProcessRuntimeEventListener<ProcessSuspendedEvent> processSuspendedListener() {
+        return processSuspendedEvent -> collectedEvents.add(processSuspendedEvent);
+    }
+
+    @Bean
+    public ProcessRuntimeEventListener<ProcessCancelledEvent> processCancelledListener() {
+        return processCancelledEvent -> collectedEvents.add(processCancelledEvent);
+    }
+
+    @Bean
+    public VariableEventListener<VariableCreatedEvent> variableCreatedEventListener() {
+        return variableCreatedEvent -> collectedEvents.add(variableCreatedEvent);
+    }
+
+    @Bean
+    @Primary
+    public VariableEventListener<VariableDeletedEvent> variableDeletedEventListener() {
+        return variableDeletedEvent -> collectedEvents.add(variableDeletedEvent);
+    }
+
+    @Bean
+    public VariableEventListener<VariableUpdatedEvent> variableUpdatedEventListener() {
+        return variableUpdatedEvent -> collectedEvents.add(variableUpdatedEvent);
+    }
+
+    @Bean
+    public TaskEventListener<TaskCreatedEvent> taskCreatedEventListener() {
+        return taskCreatedEvent -> collectedEvents.add(taskCreatedEvent);
+    }
+
+    @Bean
+    public TaskEventListener<TaskUpdatedEvent> taskUpdatedEventListener() {
+        return taskUpdatedEvent -> collectedEvents.add(taskUpdatedEvent);
+    }
+
+    @Bean
+    public TaskEventListener<TaskCompletedEvent> taskCompletedEventListener() {
+        return taskCompletedEvent -> collectedEvents.add(taskCompletedEvent);
+    }
+
+    @Bean
+    public TaskEventListener<TaskSuspendedEvent> taskSuspendedEventListener() {
+        return taskSuspendedEvent -> collectedEvents.add(taskSuspendedEvent);
+    }
+
+    @Bean
+    public TaskEventListener<TaskAssignedEvent> taskAssignedEventListener() {
+        return taskAssignedEvent -> collectedEvents.add(taskAssignedEvent);
+    }
+    
+    @Bean
+    public TaskEventListener<TaskCancelledEvent> taskCancelledEventListener() {
+        return taskCancelledEvent -> collectedEvents.add(taskCancelledEvent);
+    }
+    
+    
+    @Bean
+    public BPMNElementEventListener<BPMNSignalReceivedEvent> bpmnSignalReceivedListener() {
+        return bpmnSignalReceivedEvent -> collectedEvents.add(bpmnSignalReceivedEvent);
     }
 
 }
